Moved bottom sheet to dialogutil
