+++ core/src/main/res/layout/layout_simple_list_bottom_sheet.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:gravity="center_horizontal"
@@ -14,7 +15,7 @@
         android:backgroundTint="@color/gray_200" />
 
     <ListView
-        android:id="@+id/spinner_bottom_sheet_list_view"
+        android:id="@+id/simple_list_bottom_sheet_list_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginTop="16dp" />
